Don't require Robolectric runtime dependencies
Summary: buck1 does not enforce one of these attributes to be set. If neither are set then the `-Drobolectric.dependency.dir` flag is not included in the test run command. The consequence of this is that the tests cannot use `RobolectricTestRunner` and load "real" Android classes from the `android-all` jars -- tests which don't set these attributes seem to be regular-old JUnit style tests which don't rely on Robolectric to load Android classes. In that sense they could use the plain `java_test` rule but that's besides the point.

Reviewed By: IanChilds

Differential Revision: D44186685

fbshipit-source-id: 2d36f0ee62cf9f4814385875fff2c1f723a586fd