buck2: soft error: give back the error if you don't fire it
Summary:
Hold up, this makes no sense: if you want the error back that's probably to
throw it but then why make it a soft error.

But it does! I'd like to use soft errors as way to tag invocations that failed
in a specific way. This is actually a pretty good fit, considering:

- The failures are things we actually want to drive to zero, like a regular
  soft error.
- There is a bunch of stuff we build around soft errors that I'd
  like to use, like logging them to the event log and Scuba or rate limiting
  them.

I use this in the next diff.

Reviewed By: stepancheg

Differential Revision: D44170710

fbshipit-source-id: 81348a5012bb2ca57361cbf1e4119a05c692195d