buck2: local executor: check for missing files after a failure
Summary:
If a command fails, check whether the inputs are where they should be. If they
aren't, then fire a soft error, which the rest of this stack will translate
into something that gets logged to Ingress.

Note that this doesn't actually check deep into directories. This should do
for now.

Reviewed By: stepancheg

Differential Revision: D44170711

fbshipit-source-id: 6912befba2faa6c9695b1cc0a307f4abdd9d2520