PatternType::split -> split_providers_name
Summary:
TL;DR: code cleanup, no functional changes.

Consider we are parsing string

```
foo//bar:baz[Qux]
```

as target pattern (not as provider pattern).

Before this diff, this was parsed as package `foo/bar` and target name `baz[Qux]`, and then target name validation failed because of `[` in target name.

This diff now always parses `baz[Qux]` as a target name plus provider name, and then it fails at creation of `TargetName as PatternType` when provider is specified.

This should be more correct: we are now logically use the same grammar for target pattern and provider pattern.

Differential Revision: D44207041

fbshipit-source-id: 5941f6479a4871136f9ff45f18e9697b2133f3b7