deferred-materializer: calculate and store directory size in artifact tree, sqlite
Summary:
To avoid reading the filesystem to get tracked directory sizes during clean --stale.

It's not totally necessary to store both `digest_size` and `directory_size` and it would be simpler to replace the size in TrackedFileDigest with the directory size or store it inside ArtifactMetadata, but this way there's no need to recalculate the directory size in order to match digest against an existing artifact, and since the TrackedFileDigest is `duped` it's likely more expensive to create a new digest with only the directory size changed.

Renaming `entry_size` db field to `digest_size` for clarity (contents are the same).

Reviewed By: krallin

Differential Revision: D44049741

fbshipit-source-id: d43b53cd89571e009899681c3b859566e77bb718