Make Evaluator impl Drop
Summary:
This forces users to be more explicit about the lifetime of the
evaluator (i.e. they can't move out of objects that the evaluator has
references to without doing explicitly outside of the lifetime of the
Evaluator). This makes it so that we can more freely change the Evaluator (and
specifically allow it to capture things that impl Drop) without needing
to change client code.

Reviewed By: krallin

Differential Revision: D43941402

fbshipit-source-id: 4a78b3a921015c06e9302d1af6ad63108c02b9dc