init: Add --git flag
Summary:
Got feedback that it might not be the best to just assume everyone uses `git`. So add a new `--git` flag to do what we do now.

If a user doesn't specify `--git`,  print a message so that the user knows where to pull the buck2-prelude from.

buck2-oss is broken due to D44283648, which is being reverted in D44352395

Reviewed By: bobyangyf

Differential Revision: D44350188

fbshipit-source-id: 5b52f0ed0b49121c98389487445a9197b301330e