buck2_client_ctx: relax timed_list tests
Summary:
To render the time `Instant::elapsed()` is called, e.g. in `draw_root_first_child()`. To calculate elapsed time it calls `Instant::now()` again. And it returns time greater than `tick.start_time` in the test.

On my local Windows machine this difference is about 0.7-0.9ms.

But on Windows CI this difference can exceed 50ms resulting in `5.1s` in the rendered superconsole instead of `5.0s`.

With this change it will pass if difference is below 100ms.

Also get rid of `SystemTime::now()` to not think if it needs to be synchronized with `Instant::now()`.

Reviewed By: krallin

Differential Revision: D44426755

fbshipit-source-id: 128e4f03dc4f441efd14df0cead9aebdd08ee58d