buck2_critical_path: add support for computing critical path potentials
Summary:
This implements the algorithm from MGwynne's and gowrisurya's paper "[Longest
Path Graph: data structure + algorithm for computing max savings in workflow
graphs](https://fburl.com/gdrive/0t8pswwp)" (with trivial alterations).

The paper explains the algorithm in detail, but the intuition is that we
compute a "longest path graph" that gives us the cost of the longest path
through each node of the graph.

We then simulate reductions in critical path by checking, for every node on the
critical path, what is the cost of the longest path through any other vertex
that does *not* go through this node (meaning its runtime is not affected by
a reduction in the runtime of the critical path node).

Reviewed By: stepancheg

Differential Revision: D44380635

fbshipit-source-id: 09b0d96fba5dd721d8e3bbd58e2f6f9eb0451921