put a limit of number of processes we attempt to kill
Summary:
Something is going wrong when we gather the list of processes when running `buck2 kill` and `buck2 clean` (which does the same thing as kill plus a bit more)

## What I'm observing
1. At the end of each buck2 e2e test we run `buck2 clean`
2. Some of these `buck2 clean` processes get into a state where they each start requesting 20+GB of memory
3. The lego-windows VMs (which come with 40gb of mem) quickly run out, things slow down, tests timeout ( :'( )

## Cause?
Something about the way we walk up the process tree is wrong and can get into a bad state (don't know which yet)
Maybe the process id is being recycled? I have no clue.

## Remediation
krallin suggested limiting the number of processes we look at would be a good way to mitigate the issue.

## Pics

Reviewed By: krallin

Differential Revision: D44510718

fbshipit-source-id: 18f1473ff720c6d4f5e6fea89d76beb1695088c3