buck2: make command alias actually work properly when exec != target
Summary:
There are bugs here, namely when exec != target, we use:

- The wrong extension: we use the exec's platform and not the target's.
- The wrong variables: same problem.

Which means it totally doesn't work because now on Windows you're calling a
`.sh` script, which actually contains batch code, except for one variable name
which is Windows.

In an ideal world, we'd avoid needing to run some text replacement as a
separate command in order to fix paths in those command aliases and then none
of this would be necessary.

In a *somewhat* ideal world we'd bias the exec platform but that doesn't work
due to a bug we haven't figured out yet (see D41182948 for an unsuccessful
attempt at it).

Unfortunately we live in neither of those worlds, so let's fix it by having the
text replacement actually produce a viable script for the target platform.

Reviewed By: lmvasquezg

Differential Revision: D44533997

fbshipit-source-id: 0c938fe02afd65a465bb96a8ff5b87b6501c620b