http_archive windows support
Summary:
This was relatively straightforward. Most of http_archive is platform
agnostic. The decompress script is simple and readily adapted from bash to cmd.

One wrinkle, however, is that tar on windows is bsdtar. This bsdtar seems to just
hang when used on a zstd compressed archive. It doesn't error, it doesn't do
anything useful, it just hangs. So the decompress step needs to decompose into
manually invoking zstd -d and piping the result to tar.

Another issue is `zip` and `unzip` are not CLI commands on windows. There are powershell
commandlets for it, however, we can instead use the same bsdtar that ships
with windows to extract zip archives. It has builtin support for zip format.

Reviewed By: krallin

Differential Revision: D44514528

fbshipit-source-id: a16a7330422a2f67796e0cb26354a6d7edbd0a49