Fix the speed/age styling
Summary:
With the timed_list we use the helper duration_as_secs_elapsed which multiplies by time_speed, so if we have the time 1s, but the speed 10x, we display 10s. But we had forgot to multiply by the speed when computing the age, which is used for styling. Fix that in all the places we had it wrong.

I'm not a huge fan of where the abstraction boundaries are drawn (surely TimeSpeed should be the thing that scales a time, and we should have types to ensure we only pass scaled times around) but just improving it a bit, rather than attempting to redesign the world.

Reviewed By: bobyangyf

Differential Revision: D44597621

fbshipit-source-id: 7087ca69c6644e2092fe8b40c57fe21b48d589e9