Expose top_frame()
Summary:
For the buck debugger we want to know a simple location that each
thread is stopped at (i.e. "stopped at <file>:<line>"). For that we only
need the top frame. While there's other code currently assumes that for the
debugger it actually wants the second top frame, but that's adjusted later
in the stack (we actually want the top frame for the dap-based debugger).

Reviewed By: krallin

Differential Revision: D43941383

fbshipit-source-id: b6ab1677e949a6adc071eb9e5c3673c2adaa3cd0