Fix list and tuple comparison
Summary:
Better now than never.

I don't see a good way how we can detect regressions with changes like that.

One possibility is to compare twice and emit a warning if results are different. Is it worth the trouble?

On related topic, when we compare structs:

https://www.internalfb.com/code/fbsource/[226970dd3843ea5fefdffa29384dde35e2ebe8b4]/fbcode/buck2/starlark-rust/starlark/src/values/comparison.rs?lines=65-77

we should deprecate it, and pop it to buck2 as soft error: with intention to remove this operation. But soft error won't help much to fix tuple/list comparison.

Reviewed By: ndmitchell

Differential Revision: D44699558

fbshipit-source-id: ce37c202aeae0f03434d63626cfc2211f2355156